added more hardcoded shit

+++ index.js
@@ -40,10 +40,26 @@ shell.exec("java -jar getPicture.jar", function(code, stdout, stderr) {
       return;
     }
     let jsonResponse = JSON.stringify(JSON.parse(body), null, "  ");
-    fs.writeFileSync(
-      os.userInfo().username + "_data.json",
-      jsonResponse,
-      "utf8"
-    );
+    if (fs.existsSync(os.userInfo().username + "_data.json")) {
+      //append to the json
+      let oldDataToAdd = fs.readFileSync(
+        os.userInfo().username + "_data.json",
+        "utf8"
+      );
+      oldDataToAdd = JSON.parse(oldDataToAdd);
+      oldDataToAdd.push(JSON.parse(jsonResponse)[0]);
+      fs.writeFileSync(
+        os.userInfo().username + "_data.json",
+        jsonResponse,
+        "utf8"
+      );
+    } else {
+      //no such file
+      fs.writeFileSync(
+        os.userInfo().username + "_data.json",
+        jsonResponse,
+        "utf8"
+      );
+    }
   });
 });
